Removed the unwanted sample factor parameter from the feature importance API.
Summary:
Updates the FOS API to remove the sample factor parameter from
the feature importance function.

Using the sample factor would lead to unecessary transfer and conversion of the
entire dataset, which is undesirable. The client code should sample out the
unecessary instances on their side.

Test Plan: N/A

Reviewers: deag, joao.oliveirinha

Reviewed By: joao.oliveirinha

Differential Revision: https://review.feedzai.com/D3579